backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "The backend for the therobotoverlord project."
readme = "README.md"
authors = [
    { name = "Joshua Szepietowski", email = "joshSzep@gmail.com" },
]
requires-python = "==3.12.10"
dependencies = [
    "aerich>=0.9.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-ai-slim[logfire,openai,anthropic]>=0.2.6",
    "pydantic-graph>=0.2.6",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "tomli-w>=1.2.0",
    "tomlkit>=0.13.2",
    "tortoise-orm[accel,asyncpg]>=0.25.0",
    "uvicorn>=0.34.2",
]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpie>=3.2.4",
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pyright>=1.1.401",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.11",
    "tortoise-cli>=0.1.2",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "A",
    "C4",
    "SIM",
    "ERA",
]

[tool.ruff.lint.isort]
force-single-line = true
force-sort-within-sections = true
known-first-party = [
    "backend",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
addopts = "--cov-fail-under=1 --cov=backend --cov-report=html --cov-report=xml --cov-report=term"
testpaths = ["tests", "e2e_tests"]

[tool.mypy]
python_version = "3.12"
exclude = [
    ".venv",
    "tests",
    "e2e_tests",
]
strict = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
disallow_any_generics = true
follow_untyped_imports = true

[tool.pyright]
pythonVersion = "3.12"
venvPath = ".venv"
exclude = [
    "**/.venv",
    "**/tests",
    "**/e2e_tests",
]
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
reportPrivateImportUsage = true
reportUnknownMemberType = false
reportUnknownArgumentType = true
reportUnknownParameterType = true
reportUnknownVariableType = false
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportUntypedBaseClass = true
reportUntypedNamedTuple = true
reportFunctionMemberAccess = true
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true

[tool.aerich]
tortoise_orm = "backend.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
